requirements.txt
discord.py
SQLAlchemy
python-dotenv
tabulate
pandas